update APIs file
