+++ src/index.js
@@ -3,4 +3,3 @@ import './modules/homepage.js';
 import './modules/comment_handler.js';
 import './modules/comments-button_handler.js';
 import './modules/comments-pop-up-close-button_handler.js';
-import './modules/comments-pop-up-comments-section_handler.js';
